Complete refactor as multi-page site (#71)
* Added FAQ #32, anchor link and FAQ content

Added FAQ #32: "How do I perform a password reset on my Gameboard account?" along with the corresponding anchor link in the FAQ list.

* Remove references to apt, pip from FAQ #28

Removed the following from FAQ number 28.  This is to help clarify and avoid confusion about the fact that users are not able to install additional tools in secondary VM's -- only the default/root Kali VM.

"Mirrors of standard application repositories (Apt and Pip) are provided for use inside of the challenge environment to install additional tools on the provided virtual machines."

* Complete refactor as multi-page site

Complete overhaul and refactoring of the frontend President's Cup website to work as a multi-page bootstrap website instead of a single page website.  All content remains from the end of the PC6 competition, pending new branding and content availability for PC7.

---------

Co-authored-by: Daniel Valencia <valenciad@opstechalliance.com>