v5.0u
After several intensive days (and nights) of work, we have successfully completed all major recreation, refactoring, and critical testing ahead of schedule.

The most significant update to the codebase is the restructuring of pages to use their respective vanity URLs (e.g., /workshops instead of workshops.html). This change in particular involved changing ~100 lines in the code base, so I will be doing a lot of testing in the coming days.  Additionally, the homepage has been fully redesigned in alignment with the new layout and design.  Finally, a number of minor fixes and refinements have been implemented across the site as part of this release.  We’ve also added specific tags to instruct Cloudflare not to enforce lazy loading on certain assets—this adjustment helps address inconsistencies observed in WebKit/Safari with SVGs.