v5.4ux
This release attempts to address a server-side issue, perhaps with Cloudflare Mirage or another server side feature, that is attempting to rewrite source code at runtime and causing some issues loading some images.  To do this, there is a script that identifies and mitigates any image that is missing the src attribute.