v5.4ux (#75)
* Added FAQ #32, anchor link and FAQ content

Added FAQ #32: "How do I perform a password reset on my Gameboard account?" along with the corresponding anchor link in the FAQ list.

* Remove references to apt, pip from FAQ #28

Removed the following from FAQ number 28.  This is to help clarify and avoid confusion about the fact that users are not able to install additional tools in secondary VM's -- only the default/root Kali VM.

"Mirrors of standard application repositories (Apt and Pip) are provided for use inside of the challenge environment to install additional tools on the provided virtual machines."

* Complete refactor as multi-page site

Complete overhaul and refactoring of the frontend President's Cup website to work as a multi-page bootstrap website instead of a single page website.  All content remains from the end of the PC6 competition, pending new branding and content availability for PC7.

* v4.3

* v4.7

* v5.0u

After several intensive days (and nights) of work, we have successfully completed all major recreation, refactoring, and critical testing ahead of schedule.

The most significant update to the codebase is the restructuring of pages to use their respective vanity URLs (e.g., /workshops instead of workshops.html). This change in particular involved changing ~100 lines in the code base, so I will be doing a lot of testing in the coming days.  Additionally, the homepage has been fully redesigned in alignment with the new layout and design.  Finally, a number of minor fixes and refinements have been implemented across the site as part of this release.  We’ve also added specific tags to instruct Cloudflare not to enforce lazy loading on certain assets—this adjustment helps address inconsistencies observed in WebKit/Safari with SVGs.

* v5.1u

Now with CSS cache busting action.

/css/style.css?v5.1u

* v5.4ux

This release attempts to address a server-side issue, perhaps with Cloudflare Mirage or another server side feature, that is attempting to rewrite source code at runtime and causing some issues loading some images.  To do this, there is a script that identifies and mitigates any image that is missing the src attribute.

---------

Co-authored-by: Daniel Valencia <valenciad@opstechalliance.com>